requirements.txt
streamlit
modal
whisper
openai
python-dotenv
requests